+++ src/tools/jobspy_search.py
@@ -7,8 +7,8 @@ import time
 import os
 
 
-from models import JobSearchParams
-from settings import AppConfig
+from src.models import JobSearchParams
+from src.settings import AppConfig
 
 
 logger = logging.getLogger(__name__)
